+++ Repository/UsersRepository.cs
@@ -47,7 +47,7 @@ namespace Repository
 			await Collection.UpdateOneAsync(user => user.Id == id, update);
 		}
 
-		public async Task UpdateAccountBalance(string id, double amount)
+		public async Task UpdateAccountBalance(string id, decimal amount)
 		{
 			var update = Builders<User>.Update.Inc(u => u.AccountBalance, amount);
 			await Collection.UpdateOneAsync(user => user.Id == id, update);
